pyproject.toml
[tool.poetry]
name = "wizproxy"
version = "0.4.0"
description = "A proxy for handling encrypted Wizard101 traffic"
authors = ["Valentin B. <valentin.be@protonmail.com>"]
license = "ISC"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
loguru = "^0.7.0"
pycryptodome = "^3.18.0"
scapy = "^2.5.0"
trio = "^0.22.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"

[tool.poetry.scripts]
wizproxy = "wizproxy.__main__:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
